Split AdminController according to logic of actions. #109
