+++ lib/ex_admin/themes/admin_lte2/layout.ex
@@ -10,7 +10,7 @@ defmodule ExAdmin.Theme.AdminLte2.Layout do
     # html_opts = Keyword.get(opts, :html_opts, [])
     active_class = Keyword.get(opts, :active_class, "active")
     active_class = if link_active?(conn, path), do: active_class, else: ""
-    icon = if Path.basename(path) == "dashboard" do
+    icon = if path == ExAdmin.Utils.admin_path do
       content_tag :i, "", class: "fa fa-dashboard"
     else
       content_tag :i, String.at(name, 0), class: "nav-label label label-info"
