+++ web/templates/admin_resource/admin.html.eex
@@ -1,0 +1,1 @@
+<%= @html %>
