+++ web/templates/admin_resource/destroy.js.eex
