+++ web/templates/admin_resource/index.html.eex
