+++ web/templates/admin_resource/toggle_attr.js.eex
