+++ web/views/admin_association_view.ex
@@ -1,5 +1,5 @@
 Code.ensure_compiled(ExAdmin.Web)
-defmodule ExAdmin.AssociationView do
+defmodule ExAdmin.AdminAssociationView do
   @moduledoc false
   use ExAdmin.Web, :view
 end
