+++ web/views/admin_resource_view.ex
@@ -1,0 +1,5 @@
+Code.ensure_compiled(ExAdmin.Web)
+defmodule ExAdmin.AdminResourceView do
+  @moduledoc false
+  use ExAdmin.Web, :view
+end
